Move Dummy Wifi  cluster into own File
