+++ packages/matter-node.js-examples/src/examples/DeviceNode.ts
@@ -18,13 +18,11 @@ import { CommissioningServer, MatterServer } from "@project-chip/matter-node.js"
 
 import { BleNode } from "@project-chip/matter-node-ble.js/ble";
 import { Ble } from "@project-chip/matter-node.js/ble";
-import { ClusterServer, GeneralCommissioningCluster, NetworkCommissioning } from "@project-chip/matter-node.js/cluster";
 import { OnOffLightDevice, OnOffPluginUnitDevice, logEndpoint } from "@project-chip/matter-node.js/device";
 import { Format, Level, Logger } from "@project-chip/matter-node.js/log";
 import { StorageBackendDisk, StorageManager } from "@project-chip/matter-node.js/storage";
 import { Time } from "@project-chip/matter-node.js/time";
 import {
-    ByteArray,
     commandExecutor,
     getIntParameter,
     getParameter,
@@ -33,6 +31,7 @@ import {
     singleton,
 } from "@project-chip/matter-node.js/util";
 import { DeviceTypeId, VendorId } from "@project-chip/matter.js/datatype";
+import DummyWifiNetworkCommissioningClusterServer from "./cluster/DummyWifiNetworkCommissioningClusterServer.js";
 
 const logger = Logger.get("Device");
 
@@ -209,188 +208,7 @@ class Device {
             // purposes and just "simulates" the actions to be done. In a real world implementation this would be done by
             // the device implementor based on the relevant networking stack.
             // The NetworkCommissioningCluster and all logics are described in Matter Core Specifications section 11.8
-            const firstNetworkId = new ByteArray(32);
-            commissioningServer.addRootClusterServer(
-                ClusterServer(
-                    NetworkCommissioning.Cluster.with(NetworkCommissioning.Feature.WiFiNetworkInterface),
-                    {
-                        maxNetworks: 1,
-                        interfaceEnabled: true,
-                        lastConnectErrorValue: 0,
-                        lastNetworkId: null,
-                        lastNetworkingStatus: null,
-                        networks: [{ networkId: firstNetworkId, connected: false }],
-                        scanMaxTimeSeconds: 3,
-                        connectMaxTimeSeconds: 3,
-                    },
-                    {
-                        scanNetworks: async ({
-                            request: { ssid, breadcrumb },
-                            attributes: { lastNetworkingStatus },
-                            endpoint,
-                        }) => {
-                            console.log(
-                                `---> scanNetworks called on NetworkCommissioning cluster: ${ssid?.toHex()} ${breadcrumb}`,
-                            );
-
-                            // Simulate successful scan
-                            if (breadcrumb !== undefined) {
-                                const generalCommissioningCluster =
-                                    endpoint.getClusterServer(GeneralCommissioningCluster);
-                                generalCommissioningCluster?.setBreadcrumbAttribute(breadcrumb);
-                            }
-
-                            const networkingStatus = NetworkCommissioning.NetworkCommissioningStatus.Success;
-                            lastNetworkingStatus?.setLocal(networkingStatus);
-
-                            return {
-                                networkingStatus,
-                                wiFiScanResults: [
-                                    {
-                                        security: {
-                                            unencrypted: false,
-                                            wep: false,
-                                            wpaPersonal: false,
-                                            wpa2Personal: true,
-                                            wpa3Personal: true,
-                                        },
-                                        ssid:
-                                            ssid ||
-                                            ByteArray.fromString(getParameter("ble-wifi-scan-ssid") ?? "TestSSID"), // Set a valid existing local Wi-Fi SSID here
-                                        bssid: ByteArray.fromString(
-                                            getParameter("ble-wifi-scan-bssid") ?? "00:00:00:00:00:00",
-                                        ),
-                                        channel: 1,
-                                    },
-                                ],
-                            };
-                        },
-                        addOrUpdateWiFiNetwork: async ({
-                            request: { ssid, credentials, breadcrumb },
-                            attributes: { lastNetworkingStatus, lastNetworkId },
-                            endpoint,
-                            session,
-                        }) => {
-                            console.log(
-                                `---> addOrUpdateWiFiNetwork called on NetworkCommissioning cluster: ${ssid.toHex()} ${credentials.toHex()} ${breadcrumb}`,
-                            );
-
-                            session
-                                .getContext()
-                                .assertFailSafeArmed("Failsafe timer needs to be armed to add or update networks.");
-
-                            // Simulate successful add or update
-                            if (breadcrumb !== undefined) {
-                                const generalCommissioningCluster =
-                                    endpoint.getClusterServer(GeneralCommissioningCluster);
-                                generalCommissioningCluster?.setBreadcrumbAttribute(breadcrumb);
-                            }
-
-                            const networkingStatus = NetworkCommissioning.NetworkCommissioningStatus.Success;
-                            lastNetworkingStatus?.setLocal(networkingStatus);
-                            lastNetworkId?.setLocal(firstNetworkId);
-
-                            return {
-                                networkingStatus: NetworkCommissioning.NetworkCommissioningStatus.Success,
-                                networkIndex: 0,
-                            };
-                        },
-                        removeNetwork: async ({
-                            request: { networkId, breadcrumb },
-                            attributes: { lastNetworkingStatus, lastNetworkId },
-                            endpoint,
-                            session,
-                        }) => {
-                            console.log(
-                                `---> removeNetwork called on NetworkCommissioning cluster: ${networkId.toHex()} ${breadcrumb}`,
-                            );
-
-                            session
-                                .getContext()
-                                .assertFailSafeArmed("Failsafe timer needs to be armed to add or update networks.");
-
-                            // Simulate successful add or update
-                            if (breadcrumb !== undefined) {
-                                const generalCommissioningCluster =
-                                    endpoint.getClusterServer(GeneralCommissioningCluster);
-                                generalCommissioningCluster?.setBreadcrumbAttribute(breadcrumb);
-                            }
-
-                            const networkingStatus = NetworkCommissioning.NetworkCommissioningStatus.Success;
-                            lastNetworkingStatus?.setLocal(networkingStatus);
-                            lastNetworkId?.setLocal(firstNetworkId);
-
-                            return {
-                                networkingStatus: NetworkCommissioning.NetworkCommissioningStatus.Success,
-                                networkIndex: 0,
-                            };
-                        },
-                        connectNetwork: async ({
-                            request: { networkId, breadcrumb },
-                            attributes: { lastNetworkingStatus, lastNetworkId, lastConnectErrorValue, networks },
-                            endpoint,
-                            session,
-                        }) => {
-                            console.log(
-                                `---> connectNetwork called on NetworkCommissioning cluster: ${networkId.toHex()} ${breadcrumb}`,
-                            );
-
-                            session
-                                .getContext()
-                                .assertFailSafeArmed("Failsafe timer needs to be armed to add or update networks.");
-
-                            // Simulate successful connection
-                            if (breadcrumb !== undefined) {
-                                const generalCommissioningCluster =
-                                    endpoint.getClusterServer(GeneralCommissioningCluster);
-                                generalCommissioningCluster?.setBreadcrumbAttribute(breadcrumb);
-                            }
-
-                            const networkList = networks.getLocal();
-                            networkList[0].connected = true;
-                            networks.setLocal(networkList);
-
-                            const networkingStatus = NetworkCommissioning.NetworkCommissioningStatus.Success;
-                            lastNetworkingStatus?.setLocal(networkingStatus);
-                            lastNetworkId?.setLocal(firstNetworkId);
-                            lastConnectErrorValue?.setLocal(null);
-
-                            // Announce operational in IP network
-                            const device = session.getContext();
-                            await device.startAnnouncement();
-
-                            return {
-                                networkingStatus: NetworkCommissioning.NetworkCommissioningStatus.Success,
-                                errorValue: null,
-                            };
-                        },
-                        reorderNetwork: async ({
-                            request: { networkId, networkIndex, breadcrumb },
-                            attributes: { lastNetworkingStatus },
-                            endpoint,
-                        }) => {
-                            console.log(
-                                `---> reorderNetwork called on NetworkCommissioning cluster: ${networkId.toHex()} ${networkIndex} ${breadcrumb}`,
-                            );
-
-                            // Simulate successful connection
-                            if (breadcrumb !== undefined) {
-                                const generalCommissioningCluster =
-                                    endpoint.getClusterServer(GeneralCommissioningCluster);
-                                generalCommissioningCluster?.setBreadcrumbAttribute(breadcrumb);
-                            }
-
-                            const networkingStatus = NetworkCommissioning.NetworkCommissioningStatus.Success;
-                            lastNetworkingStatus?.setLocal(networkingStatus);
-
-                            return {
-                                networkingStatus: NetworkCommissioning.NetworkCommissioningStatus.Success,
-                                networkIndex: 0,
-                            };
-                        },
-                    },
-                ),
-            );
+            commissioningServer.addRootClusterServer(DummyWifiNetworkCommissioningClusterServer);
         }
 
         commissioningServer.addDevice(onOffDevice);
